#1094 Fix for older Roslyn.
