+++ Metalama.Framework/src/Metalama.Framework.Engine/Linking/Substitution/PropertyBackingFieldReferenceSubstitution.cs
@@ -2,10 +2,10 @@
 // SharpCrafters s.r.o. licenses this file to you under either the MIT license or a proprietary license, depending on the repository from which it was obtained.
 // Refer to LICENSE.md in the repository root for complete details.
 
+#if ROSLYN_5_0_0_OR_GREATER
+
 using Metalama.Framework.Engine.Services;
-using Metalama.Framework.Engine.Utilities.Roslyn;
 using Microsoft.CodeAnalysis;
-using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
 using static Microsoft.CodeAnalysis.CSharp.SyntaxFactory;
 
@@ -42,3 +42,5 @@ internal sealed class PropertyBackingFieldReferenceSubstitution : SyntaxNodeSubs
         }
     }
 }
+
+#endif
