+++ Metalama.Framework/src/tests/Metalama.Framework.Tests.AspectTests/Tests/Aspects/Overrides/Properties/Uninlineable_BackingFieldAccess.cs
@@ -2,6 +2,12 @@
 // SharpCrafters s.r.o. licenses this file to you under either the MIT license or a proprietary license, depending on the repository from which it was obtained.
 // Refer to LICENSE.md in the repository root for complete details.
 
+#if TEST_OPTIONS
+// @RequiredConstant(ROSLYN_5_0_0_OR_GREATER)
+#endif
+
+#if ROSLYN_5_0_0_OR_GREATER
+
 using System;
 using Metalama.Framework.Advising;
 using Metalama.Framework.Aspects;
@@ -65,3 +71,5 @@ namespace Metalama.Framework.Tests.AspectTests.TestInputs.Aspects.Overrides.Prop
         }
     }
 }
+
+#endif
